post merge work, fit and finish,
